awsrequest: Fix content length detection with `ReadFileChunk`
When requests fails to determine the content length of the body,
`AWSPreparedRequest` tries to do so by `seek`ing to the end of the
data object. This fails If the data object is a `ReadFileChunk` from
`s3transfer`, because that has `seek` and `tell` methods, but does not
support the second parameter to `seek` (`whence`). `ReadFileChunk`
does provide its size via `__len__`, though, so use that before trying
`seek`.